Improve setup.py
The current version of setup.py suffer from a typical problem among scientific
python packages that prevents the package from being installable using pip.

The problem lies in the `import numpy` line. Pip needs to be able to run
`setup.py egg_info` to figure out the dependencies but this fails when
setup.py tried to import numpy, which isn't installed yet.

This patched uses a solution inspired from http://stackoverflow.com/a/21621689.

It consists in implementing a custom build_ext class that defers the need for
numpy presence to building extension time, which is after the requirements
have been installed.